fix: UI changes to make the service pages dynamic
