+++ frontend/src/components/mcp/Sidebar.tsx
@@ -11,19 +11,6 @@ interface Server {
   status: string;
 }
 
-const getServiceIcon = (name: string) => {
-  switch(name.toLowerCase()) {
-    case 'github':
-      return Github;
-    case 'google-maps':
-      return Map;
-    case 'figma':
-      return Figma;
-    default:
-      return Bot;
-  }
-};
-
 const Sidebar = ({ onServiceSelect }: SidebarProps) => {
   const [servers, setServers] = useState<Server[]>([]);
   const [loading, setLoading] = useState(true);
@@ -47,7 +34,7 @@ const Sidebar = ({ onServiceSelect }: SidebarProps) => {
   return (
     <div className="w-80 bg-gray-800/50 backdrop-blur-sm border-r border-gray-600/30 p-6 flex flex-col gap-4">
       {servers.map((server) => {
-        const IconComponent = getServiceIcon(server.name);
+        const IconComponent = Bot;
         return (
           <button
             key={server.name}
@@ -55,7 +42,7 @@ const Sidebar = ({ onServiceSelect }: SidebarProps) => {
             className="flex items-center gap-3 p-4 bg-gray-700/30 border border-gray-600/50 rounded-xl text-white hover:bg-gray-600/40 hover:border-purple-400/50 transition-all duration-200 hover:scale-105 group"
           >
             <IconComponent className="w-5 h-5 text-gray-300 group-hover:text-purple-400 transition-colors" />
-            <span className="font-medium">{server.name}</span>
+            <span className="font-medium">{server.name.charAt(0).toUpperCase() + server.name.slice(1)}</span>
           </button>
         );
       })}
